Fix UTFGrid geocache block size detection
The function is supposed to change block size if > 95 % of found
geocache blocks have different size from expected.  This comparison
needs 21 or more caches, so leave early if there are not enough
geocaches to do the comparison.

There are no other means to prevent false detection apart from this
percentage limit.  Nevertheless, it seems extremely unlikely that all
found geocaches would be on the border or otherwise cut out so that part
of the coordinate points would be unnoticed.  If Groundspeak decides to
change their UTFGrid some day, this code needs attention anyway.