Now quick load trackable list
Previously trackables were not loaded at all when parsing a cache, instead a link to the trackable search page for the cache was stored and loaded together with all trackables.

Now when loading the cache the trackable page is parsed and trackables are quick-loaded. This reduces complexity of the Cache class and makes trackables more inline with the style of the rest of the module.