Add possibility to set password retrieval command
Instead of saving passwords as plain text in the config file this
feature allows to specify a password retrieval command instead to be
used with password stores as for example GNU pass. The password
retrieval command ('password_cmd') is mutually exclusive with the plain
text 'password' key in the config file raising a KeyError if both where
givin at the same time.