Overhaul testing with Betamax (#95)
* Fix #94

* Update Geocaching and util to support Sessions for testing

* Update tests to use Betamax

* Refactor tests to do common setup for networking.

I elected *not* to record a cassette for the login setup. If I had recorded a login setup cassette, then adding new networked tests would be difficult, because instead of actually logging in, Betamax would replay the login cassette, which wouldn't give correct data. Then the test would fail.

Instead, the setup attempts to login. If if succeeds, great. If it fails, it pretends like it succeeded and continued. If tests are only being replayed (without new ones created), everything will be fine. If new tests are recorded, they will fail due to improper authentication. They would've failed anyway, but in this case they'll fail at the recording stage instead of the setup phase.

* Refactor cookie scrubbing to take advantage of response.cookies, scrub __RequestVerificationToken

In the request, some cookies are sent in the body rather than the headers, so we need to examine that for secret values.